Create FastLED Code
This Code contains the full functionality of the audio reactive LED strips.
It requires individually addressible LED strips (NEOPIXELS) to be used properly.
Microcontroller used is the Arduino Uno, but this can be used with any other uController given there are enough GPIO pins to accomodate.

*NOTE*: Circuit Diagrams and physical drawings will be provided in the future. If there's something you want to know, feel free to send a message. 

*****
NOTE: Super Important, make sure you're using a power supply with enough supply current to power the LEDs. Max power ratings should come with the LED strip you've bought. If you dont have a suitable powersupply, there are brightness functions within the FastLed Library that you can use to lower the brightness and save on needed current to operate. This should work as a fix if the LED's start blinking erratically from not having enough current.
*****